change MAX_CXS to 16

+++ include/shared/msgr/msgr.h
@@ -3,7 +3,7 @@
 
 #include "shared/serialize/message.h"
 
-typedef uint32_t msg_addr_t;
+typedef uint16_t msg_addr_t;
 
 struct msgr;
 struct msg_sender;
